Bump to version 1.1

+++ pytom/bin/cancel_batch.py
@@ -1,7 +1,13 @@
 #!/usr/bin/env pytom
 
 import os, sys
+import subprocess
 import getpass
+from shutil import which
+
+if which('squeue') is None or which('scancel') is None:
+    print('SLURM commands not available in PATH, exiting')
+    exit()
 
 if len(sys.argv[1:]) == 2:
     try:
@@ -17,14 +23,18 @@ elif len(sys.argv[1:]) == 1:
 else:
     raise Exception('invalid input parameters')
 
-#username = os.popen('whoami').read()[:-1]
 username = getpass.getuser()
 print(username, start, end)
-running = [int(line) for line in os.popen("squeue -u " + username + " | awk '{print $1}'").readlines()[1:] if line]
+running = [int(line) for line in 
+        subprocess.run(["squeue -u " + username + " | awk '{print $1}'"], 
+            shell=True, capture_output=True, text=True
+            ).stdout.splitlines()[1:] if line]
 
 jobs = list(range(start, end+1))
-start = max(min(jobs), min(running))
+if jobs and running:
+    # Make sure that both of these lists are not empty
+    start = max(min(jobs), min(running))
 
-for i in range(start, end+1):
-    print(i)
-    if i in running: os.system('scancel {}'.format(i))
+    for i in range(start, end+1):
+        print(i)
+        if i in running: os.system('scancel {}'.format(i))
